add broadcasting to all scattered objects which seems necessary in order to prevent concurrent.futures._base.CancelledError; refactor code to pull functions into different util files
