+++ lsf-config0p10/run-config.yaml
@@ -1,6 +1,6 @@
 required_settings:
-        input_path: ../test_meshes           # Path to lod 0 meshes:
-        output_path: ../test_meshes_output   # Path to write out multires meshes:
+        input_path: /groups/cosem/cosem/ackermand/meshesForWebsite/res1decimation0p10/jrc_hela-1/er_seg           # Path to lod 0 meshes:
+        output_path: /groups/cosem/cosem/ackermand/meshesForWebsite/res1decimation0p10/jrc_hela-1/test_simpler_multires/er_seg   # Path to write out multires meshes:
         num_lods: 6                         # Number of levels of detail:
         box_size: 256                       # lod 0 box size:
 
@@ -8,4 +8,4 @@ required_settings:
 optional_decimation_settings:
         skip_decimation: False         # Skip mesh decimation if meshes exist; default is false
         decimation_factor: 2           # Factor by which to decimate faces at each lod, ie factor**lod; default is 2
-        aggressiveness: 7              # Aggressiveness to be used for decimation; default is 7
+        aggressivness: 10              # Aggressiveness to be used for decimation; default is 7
