+++ lsf-config0p25_d3a10/run-config.yaml
@@ -1,0 +1,11 @@
+required_settings:
+        input_path: /groups/cosem/cosem/ackermand/meshesForWebsite/res1decimation0p25/jrc_hela-1/er_seg           # Path to lod 0 meshes:
+        output_path: /groups/cosem/cosem/ackermand/meshesForWebsite/res1decimation0p25/jrc_hela-1/test_simpler_multires/er_seg_d3a10   # Path to write out multires meshes:
+        num_lods: 6                         # Number of levels of detail:
+        box_size: 256                       # lod 0 box size:
+
+
+optional_decimation_settings:
+        skip_decimation: False         # Skip mesh decimation if meshes exist; default is false
+        decimation_factor: 3           # Factor by which to decimate faces at each lod, ie factor**lod; default is 2
+        aggressiveness: 10             # Aggressiveness to be used for decimation; default is 7
