+++ multiresolution_mesh_creator/src/create_multiresolution_meshes.py
@@ -6,46 +6,16 @@ import time
 import os
 from os import listdir
 from os.path import isfile, join, splitext
-import yaml
-from yaml.loader import SafeLoader
 import dask
+from dask.distributed import worker_client
 import pyfqmr
-from dask.distributed import Client, worker_client
-import argparse
-import getpass
-import tempfile
-import shutil
-import multiresolution_mesh_creator.util.mesh as mesh_utils
-from ..util.logging import tee_streams, Timing_Messager, print_with_datetime
-from ..util.mesh import CompressedFragment
+import multiresolution_mesh_creator.util.mesh_util as mesh_util
+import multiresolution_mesh_creator.util.io_util as io_util
+import multiresolution_mesh_creator.util.dask_util as dask_util
 from datetime import datetime
 import logging
 
 logger = logging.getLogger(__name__)
-
-
-class Fragment:
-    """Fragment class used to store and update fragment chunk vertices, faces
-    and corresponding lod 0 fragments
-    """
-    def __init__(self, vertices, faces, lod_0_fragment_pos):
-        self.vertices = vertices
-        self.faces = faces
-        self.lod_0_fragment_pos = lod_0_fragment_pos
-
-    def update_faces(self, new_faces):
-        self.faces = np.vstack((self.faces, new_faces + len(self.vertices)))
-
-    def update_vertices(self, new_vertices):
-        self.vertices = np.vstack((self.vertices, new_vertices))
-
-    def update_lod_0_fragment_pos(self, new_lod_0_fragment_pos):
-        self.lod_0_fragment_pos.append(new_lod_0_fragment_pos)
-
-    def update(self, new_vertices, new_faces, new_lod_0_fragment_pos):
-        self.update_faces(new_faces)
-        self.update_vertices(new_vertices)
-        self.update_lod_0_fragment_pos(new_lod_0_fragment_pos)
 
 
 def my_slice_faces_plane(vertices, faces, plane_normal, plane_origin):
@@ -104,8 +74,8 @@ def update_fragment_dict(dictionary, fragment_pos, vertices, faces,
         fragment.update(vertices, faces, lod_0_fragment_pos)
         dictionary[fragment_pos] = fragment
     else:
-        dictionary[fragment_pos] = Fragment(vertices, faces,
-                                            [lod_0_fragment_pos])
+        dictionary[fragment_pos] = mesh_util.Fragment(vertices, faces,
+                                                      [lod_0_fragment_pos])
 
 
 @dask.delayed
@@ -130,7 +100,8 @@ def generate_mesh_decomposition(vertices, faces, lod_0_box_size,
     Returns:
         fragments: List of `CompressedFragments` (named tuple)
     """
-
+    vertices = vertices[0][1]
+    faces = faces[0][1]
     combined_fragments_dictionary = {}
     fragments = []
 
@@ -210,7 +181,7 @@ def generate_mesh_decomposition(vertices, faces, lod_0_box_size,
             position_quantization_bits=10)
 
         if len(draco_bytes) > 12:
-            fragment = CompressedFragment(
+            fragment = mesh_util.CompressedFragment(
                 draco_bytes, np.asarray(fragment_pos), len(draco_bytes),
                 np.asarray(fragment.lod_0_fragment_pos))
             fragments.append(fragment)
@@ -239,7 +210,7 @@ def pyfqmr_decimate(input_path, output_path, id, lod, ext, decimation_factor,
         aggressiveness [`int`]: Aggressiveness for decimation
     """
 
-    vertices, faces = mesh_utils.mesh_loader(f"{input_path}/{id}{ext}")
+    vertices, faces = mesh_util.mesh_loader(f"{input_path}/{id}{ext}")
     desired_faces = max(len(faces) // (decimation_factor**lod), 4)
     mesh_simplifier = pyfqmr.Simplify()
     mesh_simplifier.setMesh(vertices, faces)
@@ -319,11 +290,11 @@ def generate_neuroglancer_multires_mesh(output_path, num_workers, id, lods,
 
         for idx, current_lod in enumerate(lods):
             if current_lod == 0:
-                vertices, faces = mesh_utils.mesh_loader(
+                vertices, faces = mesh_util.mesh_loader(
                     f"{output_path}/mesh_lods/s{current_lod}/{id}{original_ext}"
                 )
             else:
-                vertices, faces = mesh_utils.mesh_loader(
+                vertices, faces = mesh_util.mesh_loader(
                     f"{output_path}/mesh_lods/s{current_lod}/{id}.stl")
 
             current_box_size = lod_0_box_size * 2**current_lod
@@ -351,8 +322,18 @@ def generate_neuroglancer_multires_mesh(output_path, num_workers, id, lods,
             stride = np.ceil(1.0 * (end_fragment - start_fragment) /
                              num_chunks).astype(np.int)
 
-            vertices_scattered = client.scatter(vertices)
-            faces_scattered = client.scatter(faces)
+            unique_name_vertices = f"vertices_{id}_{current_lod}_{datetime.now()}"
+            unique_name_faces = f"faces_{id}_{current_lod}_{datetime.now()}"
+            # if vertices.nbytes + faces.nbytes > 100_000:
+            # scatter large ones
+            # set hash to false here in attempt to prevent this issue: https://github.com/dask/distributed/issues/4612 ?
+            vertices_to_send = client.scatter(
+                [[unique_name_vertices, vertices]], broadcast=True)
+            faces_to_send = client.scatter([[unique_name_faces, faces]],
+                                           broadcast=True)
+            # else:
+            #    vertices_to_send = vertices
+            #    faces_to_send = faces
 
             for x in range(start_fragment[0], end_fragment[0], stride[0]):
                 for y in range(start_fragment[1], end_fragment[1], stride[1]):
@@ -360,12 +341,13 @@ def generate_neuroglancer_multires_mesh(output_path, num_workers, id, lods,
                                    stride[2]):
                         current_start_fragment = np.array([x, y, z])
                         current_end_fragment = current_start_fragment + stride
+                        #unique_name_function = f"{unique_name_vertices}_{unique_name_faces}_{x}_{y}_{z}_{datetime.now()}"
                         results.append(
                             generate_mesh_decomposition(
-                                vertices_scattered, faces_scattered,
+                                vertices_to_send, faces_to_send,
                                 lod_0_box_size, current_start_fragment,
                                 current_end_fragment, current_lod, num_chunks))
-
+            client.rebalance()
             dask_results = dask.compute(*results)
 
             # Remove empty slabs
@@ -380,12 +362,12 @@ def generate_neuroglancer_multires_mesh(output_path, num_workers, id, lods,
 
             results = []
             dask_results = []
-            mesh_utils.write_mesh_files(
+            mesh_util.write_mesh_files(
                 f"{output_path}/multires", f"{id}", fragments, current_lod,
                 lods[:idx + 1],
                 np.asarray([lod_0_box_size, lod_0_box_size, lod_0_box_size]))
 
-    print_with_datetime(
+    io_util.print_with_datetime(
         f"Completed creation of multiresolution neuroglancer mesh for mesh {id}!",
         logger)
 
@@ -411,137 +393,16 @@ def generate_all_neuroglancer_multires_meshes(output_path, num_workers, ids,
             generate_neuroglancer_multires_mesh(output_path, num_workers, id,
                                                 lods, original_ext,
                                                 lod_0_box_size))
-
     dask.compute(*results)
-
-
-def _set_local_directory(cluster_type):
-    # from https://github.com/janelia-flyem/flyemflows/blob/6ae20cb58fd55e74b47a43efdab7e09908a346ba/flyemflows/util/dask_util.py
-    # This specifies where dask workers will dump cached data
-    local_dir = dask.config.get(f"jobqueue.{cluster_type}.local-directory",
-                                None)
-    if local_dir:
-        return
-
-    user = getpass.getuser()
-    local_dir = None
-    for d in [f"/scratch/{user}", f"/tmp/{user}"]:
-        try:
-            os.makedirs(d, exist_ok=True)
-        except OSError:
-            continue
-        else:
-            local_dir = d
-            dask.config.set(
-                {f"jobqueue.{cluster_type}.local-directory": local_dir})
-
-            # Set tempdir, too.
-            tempfile.tempdir = local_dir
-
-            # Forked processes will use this for tempfile.tempdir
-            os.environ['TMPDIR'] = local_dir
-            break
-
-    if local_dir is None:
-        raise RuntimeError(
-            "Could not create a local-directory in any of the standard places."
-        )
-
-
-def start_dask(num_workers):
-
-    # Update dask
-    with open("dask-config.yaml") as f:
-        config = yaml.load(f, Loader=SafeLoader)
-        dask.config.update(dask.config.config, config)
-
-    cluster_type = next(iter(dask.config.config['jobqueue']))
-    _set_local_directory(cluster_type)
-
-    if cluster_type == 'local':
-        from dask.distributed import LocalCluster
-        cluster = LocalCluster(n_workers=num_workers, threads_per_worker=1)
-    else:
-        if cluster_type == 'lsf':
-            from dask_jobqueue import LSFCluster
-            cluster = LSFCluster()
-        elif cluster_type == 'slurm':
-            from dask_jobqueue import SLURMCluster
-            cluster = SLURMCluster()
-        elif cluster_type == 'sge':
-            from dask_jobqueue import SGECluster
-            cluster = SGECluster()
-        cluster.scale(num_workers)
-
-    client = Client(cluster)
-    # while (client.status == "running"
-    #        and len(cluster.scheduler.workers) < num_workers):
-    #     print(
-    #         f"Waiting for {num_workers - len(cluster.scheduler.workers)} workers..."
-    #     )
-
-    return client
-
-
-def read_run_config(config_path):
-    with open(f"{config_path}/run-config.yaml") as f:
-        config = yaml.load(f, Loader=SafeLoader)
-        required_settings = config['required_settings']
-        optional_decimation_settings = config.get(
-            "optional_decimation_settings", {})
-
-        if "skip_decimation" not in optional_decimation_settings:
-            optional_decimation_settings["skip_decimation"] = False
-        if "decimation_factor" not in optional_decimation_settings:
-            optional_decimation_settings["decimation_factor"] = 2
-        if "aggressiveness" not in optional_decimation_settings:
-            optional_decimation_settings["aggressiveness"] = 7
-
-        return required_settings, optional_decimation_settings
-
-
-def parser_params():
-    parser = argparse.ArgumentParser(
-        description=
-        'Code to convert single-scale (or a set of multi-scale) meshes to the neuroglancer multi-resolution mesh format'
-    )
-    parser.add_argument(
-        "config_path",
-        type=str,
-        help="Path to directory containing run-config.yaml and dask-config.yaml"
-    )
-    parser.add_argument(
-        '--num-workers',
-        '-n',
-        type=int,
-        default=1,
-        help=
-        'Number of workers to launch (i.e. each worker is launched with a single bsub command)'
-    )
-
-    return parser.parse_args()
-
-
-def setup_execution_directory(config_path):
-    # Create execution dir (copy of template dir) and make it the CWD
-    # from flyemflows: https://github.com/janelia-flyem/flyemflows/blob/03cfd79ccc1dcd4903007b36759f4b677ca5c67e/flyemflows/bin/launchflow.py
-    timestamp = f'{datetime.now():%Y%m%d.%H%M%S}'
-    execution_dir = f'{config_path}-{timestamp}'
-    execution_dir = os.path.abspath(execution_dir)
-    shutil.copytree(config_path, execution_dir, symlinks=True)
-    os.chmod(f'{execution_dir}/run-config.yaml', 0o444)  # read-only
-    print_with_datetime(f"Setup working directory as {execution_dir}.", logger)
-
-    return execution_dir
 
 
 def main():
     submission_directory = os.getcwd()
 
     # If more than 1 thread per worker, run into issues with decimation?
-    args = parser_params()
+    args = io_util.parser_params()
     num_workers = args.num_workers
-    required_settings, optional_decimation_settings = read_run_config(
+    required_settings, optional_decimation_settings = io_util.read_run_config(
         args.config_path)
 
     input_path = required_settings['input_path']
@@ -553,10 +414,11 @@ def main():
     decimation_factor = optional_decimation_settings['decimation_factor']
     aggressiveness = optional_decimation_settings['aggressiveness']
 
-    execution_directory = setup_execution_directory(args.config_path)
+    execution_directory = dask_util.setup_execution_directory(
+        args.config_path, logger)
     logpath = f'{execution_directory}/output.log'
 
-    with tee_streams(logpath):
+    with io_util.tee_streams(logpath):
 
         try:
             os.chdir(execution_directory)
@@ -570,34 +432,36 @@ def main():
 
             t0 = time.time()
 
-            # Start dask
-            with Timing_Messager("Starting dask cluster", logger):
-                client = start_dask(num_workers)
-            print_with_datetime(
-                f"Check {client.cluster.dashboard_link} for status", logger)
             # Mesh decimation
             if not skip_decimation:
-                with Timing_Messager("Generating decimated meshes", logger):
-                    generate_decimated_meshes(input_path, output_path, lods,
-                                              mesh_ids, mesh_ext,
-                                              decimation_factor,
-                                              aggressiveness)
+                # Start dask
+                with dask_util.start_dask(num_workers, "decimation", logger):
+                    with io_util.Timing_Messager("Generating decimated meshes",
+                                                 logger):
+                        generate_decimated_meshes(input_path, output_path,
+                                                  lods, mesh_ids, mesh_ext,
+                                                  decimation_factor,
+                                                  aggressiveness)
 
-            # Create multiresolution meshes
-            with Timing_Messager("Generating multires meshes", logger):
-                generate_all_neuroglancer_multires_meshes(
-                    output_path, num_workers, mesh_ids, lods, mesh_ext,
-                    lod_0_box_size)
+            # Restart dask to clean up cluster before multires assembly
+            with dask_util.start_dask(num_workers, "multires creation",
+                                      logger):
+                # Create multiresolution meshes
+                with io_util.Timing_Messager("Generating multires meshes",
+                                             logger):
+                    generate_all_neuroglancer_multires_meshes(
+                        output_path, num_workers, mesh_ids, lods, mesh_ext,
+                        lod_0_box_size)
 
             # Writing out top-level files
-            with Timing_Messager("Writing info and segment properties files",
-                                 logger):
+            with io_util.Timing_Messager(
+                    "Writing info and segment properties files", logger):
                 output_path = f"{output_path}/multires"
-                mesh_utils.write_segment_properties_file(output_path)
-                mesh_utils.write_info_file(output_path)
+                mesh_util.write_segment_properties_file(output_path)
+                mesh_util.write_info_file(output_path)
 
-            print_with_datetime(f"Complete! Elapsed time: {time.time() - t0}",
-                                logger)
+            io_util.print_with_datetime(
+                f"Complete! Elapsed time: {time.time() - t0}", logger)
         finally:
             os.chdir(submission_directory)
